int -> short for column positions
mysql has a hard-limit of 4096 columns so we're wasting a few bytes here.

+++ src/main/java/com/zendesk/maxwell/schema/MysqlSavedSchema.java
@@ -463,7 +463,7 @@ public class MysqlSavedSchema {
 
 		Database currentDatabase = null;
 		Table currentTable = null;
-		int columnIndex = 0;
+		short columnIndex = 0;
 
 		while (rs.next()) {
 			// Database
