+++ src/main/java/com/zendesk/maxwell/schema/columndef/EnumeratedColumnDef.java
@@ -6,7 +6,7 @@ abstract public class EnumeratedColumnDef extends ColumnDef  {
 	@JsonProperty("enum-values")
 	protected String[] enumValues;
 
-	public EnumeratedColumnDef(String name, String type, int pos, String [] enumValues) {
+	public EnumeratedColumnDef(String name, String type, short pos, String [] enumValues) {
 		super(name, type, pos);
 		this.enumValues = enumValues;
 	}
